windows_exe_app/requirements-windows-exe.txt
openai-whisper
gradio
pyinstaller
